Users can now add a picture when they register
