+++ libs/api/donate/api/feature/src/lib/api-donate-api-feature.module.ts
@@ -5,11 +5,9 @@ import { DonateResolver} from './donate-resolver'
 import { DonateRepository } from '@charity-spot/api/donate/repository/data-access';
 import { PrismaService } from '@charity-spot/api/shared/services/prisma';
 
-import { FirebaseService } from '@charity-spot/api/shared/services/prisma';
-
 @Module({
   controllers: [],
   imports: [ApiDonateServiceFeatureModule],
-  providers: [DonateResolver, DonateService, DonateRepository, PrismaService, FirebaseService]
+  providers: [DonateResolver, DonateService, DonateRepository, PrismaService]
 })
 export class ApiDonateApiFeatureModule {}
