+++ libs/api/shared/auth/src/lib/api-shared-auth.ts
@@ -66,16 +66,3 @@ export function apiSharedAuth(): string {
     export async function compareCodes(internal: number, external: number) {
       return internal === external;
     }
-
-
-//random
-  //generator
-    export async function randomStringGenerator() {
-      let text = "";
-      const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-    
-      for (let i = 0; i < 10; i++)
-        text += alphabet.charAt(Math.floor(Math.random() * alphabet.length));
-    
-      return text;
-    }
