+++ libs/api/shared/services/prisma/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/shared-prisma.service';
-export {FirebaseService} from './lib/FirebaseRepository.repository';
